requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Image processing
opencv-python>=4.8.0
pillow>=10.0.0
scikit-image>=0.21.0

# Visualization
matplotlib>=3.7.0

# Google Earth Engine
earthengine-api>=0.1.370
geemap>=0.28.0

# Web UI
streamlit>=1.28.0
gradio>=4.0.0

# Metrics
pytorch-msssim>=1.0.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
pyyaml>=6.0

# Notebook support
ipykernel>=6.25.0
jupyter>=1.0.0

# Testing
pytest>=7.4.0
